src/story32/RecruiterAssistantChatbot.java
package story32;

import java.util.Scanner;

public class RecruiterAssistantChatbot {

    // Assume a CareerGoalsAdvisor class for providing guidance on setting realistic career goals
    static class CareerGoalsAdvisor {
        public void provideCareerGoalsGuidance() {
            // Simulate providing guidance on setting realistic career goals
            System.out.println("Here is some guidance on setting realistic career goals as a job seeker:");
            System.out.println("1. Assess your skills, strengths, and interests to identify areas of focus.");
            System.out.println("2. Define short-term and long-term career objectives based on your assessment.");
            System.out.println("3. Ensure your goals are specific, measurable, achievable, relevant, and time-bound (SMART).");
            System.out.println("4. Research industry trends and job market demands to align your goals with opportunities.");
            System.out.println("5. Seek mentorship and advice from professionals in your desired field.");
            // Add more personalized guidance based on your specific recommendations
        }
    }

    public static void main(String[] args) {
        Scanner scanner = new Scanner(System.in);

        System.out.println("Welcome to the Recruiter Assistant Chatbot!");

        // Assume a CareerGoalsAdvisor instance for providing career goals guidance
        CareerGoalsAdvisor goalsAdvisor = new CareerGoalsAdvisor();

        do {
            System.out.println("As a job seeker, how can I assist you with setting realistic career goals?");
            System.out.println("1. Provide guidance on setting realistic career goals.");
            System.out.println("2. Exit");

            int userChoice = Integer.parseInt(scanner.nextLine());

            switch (userChoice) {
                case 1:
                    // Provide guidance on setting realistic career goals
                    goalsAdvisor.provideCareerGoalsGuidance();
                    break;
                case 2:
                    System.out.println("Exiting Recruiter Assistant Chatbot. Goodbye!");
                    scanner.close();
                    System.exit(0);
                    break;
                default:
                    System.out.println("Invalid choice. Please enter a valid option.");
                    break;
            }

            System.out.println("Do you want assistance with anything else? (yes/no)");
        } while (scanner.nextLine().equalsIgnoreCase("yes"));

        System.out.println("Exiting Recruiter Assistant Chatbot. Goodbye!");
        scanner.close();
    }
}
